requirements.txt
ultralytics
numpy
matplotlib
pandas
seaborn
ipykernel
streamlit
Pillow